fix: correcting remaining testings due to refactoring
