+++ base/infrastructure/file_management/file_writer/json_file_writer.py
@@ -9,7 +9,8 @@ from base.infrastructure.file_management.file_handler import FileHandler
 
 # Domain
 from base.domain.common.value_objects import DictValueObject
-from base.domain.file_management.file_constants import file_mode_values
+from base.domain.file_management.file_constants.file_mode_values import file_mode_values
+from base.domain.file_management.file_constants.file_type_values import file_type_values
 from base.domain.file_management.file_handler import BaseFileHandler
 from base.domain.file_management.file_writer import BaseFileWriter
 from base.domain.path_management.path_doubles import BasePath
@@ -29,7 +30,13 @@ class JsonFileWriter(BaseFileWriter):
             raise ValueError(f"Error path_obj: {path_obj} is not an instance of {BasePath}")
 
         if not isinstance(file_handler, (BaseFileHandler, type(None))):
-            raise ValueError(f"Error path_obj: {path_obj} is not an instance of {BasePath}")
+            raise ValueError(f"Error file_handler: {file_handler} is not an instance of {BaseFileHandler}")
+
+        if not path_obj.exists():
+            raise ValueError(f"Error path_obj: {path_obj} doesn't exists in file system")
+
+        if not path_obj.suffix == file_type_values.json:
+            raise ValueError(f"Error path_obj.suffix: {path_obj.suffix} it's not json type")
 
         self.__path_obj = path_obj
         self.__file_handler = file_handler or FileHandler(path_obj=self.__path_obj, file_mode=file_mode_values.write)
