+++ tests/base/infrastructure/file_management/file_writer/test_yaml_file_writer.py
@@ -37,13 +37,13 @@ def test_yaml_file_writer_validation_with_valid_params():
 
     target_path = "/home/user1/project1/folder1"
     target_path = f"{target_path}/{fake_file.name}{fake_file.suffix}"
+
     fake_path = PathFaker(target_path=target_path, target_path_type=path_types_values.file, fake_file=fake_file)
     fake_path.touch()
 
-    path_handler = PathItemCreator(path_obj=fake_path)
-    file_handler = FileHandler(path_handler=path_handler, file_obj=fake_file)
+    file_handler = FileHandler(path_obj=fake_path, file_obj=fake_file)
 
-    yaml_file_writer = YamlFileWriter(path_handler=path_handler, file_handler=file_handler)
+    yaml_file_writer = YamlFileWriter(path_obj=fake_path, file_handler=file_handler)
     yaml_file_writer.write_file(data=data)
 
     assert fake_file.content == expected_content
@@ -59,15 +59,14 @@ def test_yaml_file_writer_validation_with_invalid_path_not_exists():
 
     target_path = "/home/user1/project1/folder1"
     target_path = f"{target_path}/{fake_file.name}{fake_file.suffix}"
+
     fake_path = PathFaker(target_path=target_path, target_path_type=path_types_values.file, fake_file=fake_file)
+    file_handler = FileHandler(path_obj=fake_path, file_obj=fake_file)
 
-    path_handler = PathItemCreator(path_obj=fake_path)
-
-    expected_error_message = f"Error stored_path: {path_handler.stored_path} doesn't exists in file system"
+    expected_error_message = f"Error path_obj: {fake_path} doesn't exists in file system"
 
     with pytest.raises(ValueError) as err:
-        file_handler = FileHandler(path_handler=path_handler, file_obj=fake_file)
-        yaml_file_writer = YamlFileWriter(path_handler=path_handler, file_handler=file_handler)
+        yaml_file_writer = YamlFileWriter(path_obj=fake_path, file_handler=file_handler)
 
     assert err.value.args[0] == expected_error_message
 
@@ -82,16 +81,16 @@ def test_yaml_file_writer_validation_with_invalid_file_type():
 
     target_path = "/home/user1/project1/folder1"
     target_path = f"{target_path}/{fake_file.name}{fake_file.suffix}"
+
     fake_path = PathFaker(target_path=target_path, target_path_type=path_types_values.file, fake_file=fake_file)
     fake_path.touch()
 
-    path_handler = PathItemCreator(path_obj=fake_path)
-    file_handler = FileHandler(path_handler=path_handler, file_obj=fake_file)
+    file_handler = FileHandler(path_obj=fake_path, file_obj=fake_file)
 
-    expected_error_message = f"Error path_handler: {path_handler} file_type_suffix is not .yaml"
+    expected_error_message = f"Error path_obj.suffix: {fake_path.suffix} it's not yaml type"
 
     with pytest.raises(ValueError) as err:
-        yaml_file_writer = YamlFileWriter(path_handler=path_handler, file_handler=file_handler)
+        yaml_file_writer = YamlFileWriter(path_obj=fake_path, file_handler=file_handler)
 
     assert err.value.args[0] == expected_error_message
